wip rework

+++ MachineGun.java
@@ -38,7 +38,7 @@ public class MachineGun extends Weapon{
      * @param target what to shoot at
      * @param image the image for the gun
      */
-    public MachineGun(Vector target, GreenfootImage image, int maxShotDeviance, Car owner, int attackSpeed, double damage){
+    public MachineGun(Vector target, GreenfootImage image, int maxShotDeviance, Car owner, int attackSpeed, int damage){
         super(target, image, owner, attackSpeed, damage);
         this.maxShotDeviance = maxShotDeviance;
         r = new Random();
@@ -54,7 +54,7 @@ public class MachineGun extends Weapon{
      * @param target what to shoot at
      * @param image the image for the gun
      */
-    public MachineGun(Vector target, GreenfootImage image, int maxShotDeviance, Car owner, int attackSpeed, double damage, double bulletSpeed){
+    public MachineGun(Vector target, GreenfootImage image, int maxShotDeviance, Car owner, int attackSpeed, int damage, double bulletSpeed){
         super(target, image, owner, attackSpeed, damage);
         this.maxShotDeviance = maxShotDeviance;
         r = new Random();
